add logic to allow dynamic clear / preloading of timelines
